Fixed to build more easily

+++ external/board/edgeos/rpi4-boot/post-build.sh
@@ -12,11 +12,15 @@ install -D -m 0644 ${BR2_EXTERNAL_EDGEOS_PATH}/board/edgeos/rpi4-boot/cmdline.tx
 install -D -m 0644 edgeos-version ${BINARIES_DIR}/edgeos-version   # On FAT partition, see genimage
 install -D -m 0644 edgeos-version ${TARGET_DIR}/etc/edgeos-version # In CPIO initramfs file
 
-# Create archive of kernel modules, to later be added to the rootfs
-tar -czvf ${BINARIES_DIR}/modules.tar.gz -C ${TARGET_DIR}/lib/modules . > /dev/null 2>&1
+if [ -d "${TARGET_DIR}/lib/modules" ]; then
 
-# Reduce size of initramfs
-rm -rf ${TARGET_DIR}/lib/modules
+    # Create archive of kernel modules, to later be added to the rootfs
+    tar -czvf ${BINARIES_DIR}/modules.tar.gz -C ${TARGET_DIR}/lib/modules . > /dev/null 2>&1
+
+    # Reduce size of initramfs
+    rm -rf ${TARGET_DIR}/lib/modules
+
+fi
 
 # Remove unused overlays
 find ${BINARIES_DIR}/rpi-firmware/overlays -type f -not -name 'miniuart-bt.dtbo' -delete
